news/pyproject.toml
[project]
name = "news-neo4j-importer"
version = "0.1.0"
description = "NYT News to Neo4j knowledge graph importer with AI embeddings"
authors = [
    {name = "News Import Project", email = ""}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.1",
    "pandas>=1.3.0",
    "python-dotenv>=1.0.0",
    "neo4j>=5.0.0",
    "numpy>=1.21.0",
    "tqdm>=4.62.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
news-import = "news_import:main"
news-validate = "test_news_data:main"
news-query = "sample_queries:main"
news-embeddings = "news_embeddings:main"
news-vector-search = "vector_search:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
